feat: Redis Client module 분리
